Move the provider `Domain` class into `Providers`
Eventually the `Domains` module is going to be removed. Instead the
`Domain` base class will live directly within the `Providers` module.

The `by_provider` method is no longer required as it has been replaced
with the `Providers.select` method
